% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add1.R
\name{add1CA}
\alias{add1CA}
\alias{N_add1CA}
\alias{add1Column}
\title{function to add a columns}
\usage{
add1CA(t, k, v, ...)

N_add1CA(t, k, v)

add1Column(D, t, v, ...)
}
\arguments{
\item{t}{integer: the required strength, or the strength of \code{D} (not checked)}

\item{k}{integer: the required number of columns}

\item{v}{integer: the required number of levels, or the number of levels of \code{D}}

\item{...}{currently not used}

\item{D}{a strength \code{t} CA with \code{v} levels in \code{k-1} columns\cr
(integer levels, starting at 0 or 1)}
}
\value{
\code{add1Column} and \code{add1CA} return a CA of strength \code{t} in \code{v} levels in
\code{k} columns.\cr
It will have at most \code{\link{bestN}(t-2,k-2,v)*v*(v-1)} additional rows.\cr
\code{N_add1CA} returns the number of runs obtainable by \code{add1CA}.
}
\description{
increases the number of columns by 1,
adding at most bestN(t-2,k-1,v)\emph{v}(v-1) rows.
}
\section{Details}{

Function \code{add1CA} applies function \code{add1Column} the the best available
strength \code{t} CA at \code{v} levels with \code{k-1} columns in the role of \code{D}.\cr
The construction relies on appending to \code{D} a copy of the last column of \code{D},
followed by adding rows that consist of\cr
the best strength \code{t-2} CA at \code{v} levels with \code{k-2} columns, repeated \code{v*(v-1)} times, in the first \code{k-2} columns,\cr
each combined with one of the \code{v(v-1)} distinct level pairs for the last two columns.\cr
For \code{t=2}, the best CA is simply a constant row of flexible values.

The construction is not interesting for small CAs, but may be useful, where
good large CAs, e.g., ones constructed by cyclotomy, have one less column
than needed (and a good CA with strength \code{t-2} for \code{k-1} columns is
available).
}

\examples{
D <- powerCA(3,1331,2)
dim(D)
D2 <- add1Column(D, 3, 2)
dim(D2)
eCAN(3,1331,2)
eCAN(3,1332,2)
N_add1CA(3,1332,2)

dim(add1CA(3,1332,2))

\donttest{
A <- add1CA(6,2504,2)
dim(A)
eCAN(6,2504,2) ## much better
Ns(4,2504,2) ## its so much worse because this ingredient is poor
3047 + (1394 - 272)*2
eCAN(4,2504,2) ## Colbourn and Zhou power construction not implemented
}

}
\references{
CKRS, Theorem 3.2
}
