% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{funmakefromstrings}
\alias{funmakefromstrings}
\title{auxiliary functions}
\description{
non-visible functions to be used in other functions,
and currently also previous drafts that may be deleted
}
\details{
funmakefromstrings: read example designs from copy-pasted strings
iacheck: check a two-column matrix for correct interaction structure
subia: checks whether ia1 is subset of ia2
levels.no from DoE.base
fasttab: table replacement without using factors
ord from DoE.base
rho: extracts run numbers of D that hold the interaction ia
allints: function that extracts all interactions for (up to) t factors from a run
failcands: fail candidates from a test suite with a single failed run
failcands2: (failed) attempt to speed up failcands
rhos_for_ints
is.CA: checks for covering array properties,
can return distribution of covering frequencies or actual row sets
is.LA: checks for locating array properties
}
