% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_GFcalculus.R
\name{int2poly}
\alias{int2poly}
\title{auxiliary functions for Galois field arithmetics}
\description{
non-visible functions to be used in other functions,
including the construction for OA(q^3, 3, q+2, q) for q=2^s
}
\details{
Most functions are taken from SOAs. Function createBush3PowerOf2
implements the construction via permutation vectors as stated in Sherwood (2006)
}
