% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_ingredients.R
\name{OD2}
\alias{OD2}
\title{auxiliary ingredients}
\description{
non-visible functions for creating ingredients
}
\section{Details}{


OD2: ordered designs of strength 2, for the Sherwood mixed level construction
OD3: ordered designs of strength 3, for ordered design (CCL)
DCA: DCA according to Shokri and Moura (2025)
}

