% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_mixed.R
\name{one_to_howmany}
\alias{one_to_howmany}
\title{auxiliary functions for mixed level CAs}
\description{
non-visible functions to be used in functions for MCA creation,
and currently also previous drafts that may be deleted
key resource: Sherwood (2008) and Groemping (2025)
}
\details{
OD: ingredient creation for array expansion
one_to_howmany: determines the number of factors a column can be expanded to
colexpandOD: expands a column
makeNA: makes appropriate levels into flexible (NA) values
colexpandODr: repeat colexpandOD
find_mix_construction_separate: finds construction details for
a requested setting of levels, for use in MCA2
find_mix_construction: unfinished, attempts to be more efficient
but is way more complicated
outerseparate: workhorse function of function MCA2
}
