% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_power.R
\name{createDHF}
\alias{createDHF}
\alias{alias}
\alias{T}
\alias{SCA_Busht}
\title{Auxiliary functions for Colbourn Torres-Jimenez Power construction}
\usage{
createDHF(oa, toa, tdhf, vdhf=tdhf)

T(DHF, rs, reduce)

SCA_Busht(q, t)
}
\arguments{
\item{oa}{an orthogonal array}

\item{toa}{strength of the orthogonal array}

\item{tdhf}{requested strength of the DHF}

\item{vdhf}{number of classes of the DHF, the default yields a PHF}

\item{DHF}{a DHF or DHHF, for which the number of levels for
rows indexed by elements of \code{rs}
is to be reduced by \code{reduce}}

\item{rs}{vector of row indices whose number of levels is to be reduced}

\item{reduce}{number of levels to reduce}

\item{q}{number of levels, prime or prime power}

\item{t}{requested strength}
}
\value{
Function \code{createDHF} creates a DHF or PHF of minimum
number of rows for the combination of \code{toa}, \code{tdhf} and
\code{vdhf}.\cr
Function \code{T} reduces the number of levels for the specified rows
by the amount stated in \code{reduce}, aiming for retaining the
maximum number of columns.\cr
Function \code{SCA_Busht} is a convenience function that creates
a strength \code{t} design with \code{q} levels using function
\code{createBusht} of R package lhs.
}
\description{
These functions support the implementation of the power constructions
by Colbourn and Torres-Jimenez
}
\section{Details}{

Function \code{createDHF} creates a distributed hash family according to
Lemma 1.1 or Lemma 1.2 of Colbourn and Torres-Jimenez (2010).

Function \code{T} applies Lemma 3.1 of Colbourn and Torres-Jimenez (2010).
}

\examples{
## a DHF of strength 3 with 2 classes
## from the Bush array of strength 2
## (which coincides with the Bose array)
mydhf <- createDHF(SCA_Busht(11, 2), 2, 3, 2)
dim(mydhf)
mydhf[,1:6]
mydhf[,116:121]
## reduce the third row by 6 levels to 5 levels
mydhhf <- T(mydhf, 3, 6)
dim(mydhhf)
mydhhf[,50:55]

## a PHF of strength 4 from a strength 2 OA
myphf19 <- createDHF(SCA_Bose(19), 2, 4)
dim(myphf19)

## a DHF of strength 4 with only 2 classes
## from a strength 2 OA
mydhf19 <- createDHF(SCA_Bose(19), 2, 4, 2)
dim(mydhf19)

}
\references{
Colbourn and Torres-Jimenez (2010)
}
