% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{lower_bound_CA}
\alias{lower_bound_CA}
\alias{lower_bound_CA(t,}
\alias{k,}
\alias{v)}
\title{functions for bounds for CA sizes}
\usage{
lower_bound_CA(t, k, v)
}
\arguments{
\item{t}{requested strength of CA (positive integer number >= 2)}

\item{k}{number of columns, must be at least \code{t}}

\item{v}{number of levels for each column:
a single integer number or a vector of integer numbers of length \code{k}}
}
\value{
\code{lower_bound_CA} returns a positive number
}
\description{
The function(s) provide bounds for CA sizes.
}
\section{Details}{

Function \code{lower_bound_CA} provides the trivial lower bound which is the
largest product of \code{t} numbers of levels.
}

\examples{
lower_bound_CA(3, 75, 2)  ## 2^3
lower_bound_CA(3, 75, c(rep(2,72), 3, 5, 8)) ## 3*5*8

}
