% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca-class.R
\name{print.ca}
\alias{print.ca}
\title{Class ca and its methods}
\usage{
\method{print}{ca}(x, ...)
}
\arguments{
\item{x}{the object to be printed}

\item{\dots}{further arguments to \code{print.default}}
}
\value{
The print method invisibly returns the input objext \code{x}
}
\description{
This documentation file documents the class ca (for CAs = Covering Arrays) and its method.
}
\section{Details}{

The print method for class \code{ca} suppresses printing of most attributes,
but indicates which attributes are available.

The S3 class \code{ca} is for a matrix with attributes.
Common attributes are \code{origin}, \code{t}, \code{Call},
\code{eCAN}, \code{date}, \code{CAs-version}, \code{flexible} or \code{PCAstatus}.
Not all attributes are present for all \code{ca} objects.

The matrix elements are consecutive non-negative integers starting with 0 or 1,
and possibly missing values that stand for flexible values, i.e.,
values that can be arbitrarily chosen without deteriorating the strength of the CA.
}

