% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCyclotomyConditions.R
\name{checkcond1}
\alias{checkcond1}
\alias{checkcond2}
\alias{checkcond3}
\alias{checkcond4}
\alias{checkcond3a}
\alias{checkcond3b}
\alias{checkcond4a}
\alias{checkcond4b}
\title{Functions for checking the Cyclotomy conditions}
\usage{
checkcond1(t, v, q, iter=FALSE, progress=FALSE)

checkcond2(t, v, q, iter=FALSE, progress=FALSE)

checkcond3(t, v, q, iter=FALSE, progress=FALSE)

checkcond4(t, v, q, iter=FALSE, progress=FALSE)

checkcond3a(t, v, q, iter=FALSE, progress=FALSE, type3=TRUE)

checkcond3b(t, v, q, iter=FALSE, progress=FALSE, type3=TRUE)

checkcond4a(t, v, q, iter=FALSE, progress=FALSE, type4=TRUE)

checkcond4b(t, v, q, iter=FALSE, progress=FALSE, type4=TRUE)
}
\arguments{
\item{t}{CA strength to be checked}

\item{v}{number of levels}

\item{q}{prime or prime power}

\item{iter}{logical (default FALSE): if FALSE, all combinations are calculated at once (slightly faster); otherwise, combinations are iterated through one at a time (works for large cases where memory is not sufficient for the other strategy).}

\item{progress}{logical (default FALSE): if TRUE, show a progress bar.
For the a and b types, two progress bars will be shown (second one progresses much faster than first one).}

\item{type3}{Should the type 3 check be run? Set to FALSE, if it is already known that the type 3 condition is fulfilled.}

\item{type4}{Should the type 4 check be run? Set to FALSE, if it is already known that the type 4 condition is fulfilled.}
}
\value{
a logical indicating whether the condition is fulfilled.
}
\description{
Conditions for constructions 1 to 4 and their modifications
of Colbourn (2010) can be checked.
The checks take a long time for large cases.
}
\section{Details}{
 \code{checkcond1} to \code{checkcond4} do
the basic checks. Checks for constructions 3a and 4a consist of
checks for constructions 3 and 4, combined with checks for construction 1
for strength t-1,\cr
and checks for constructions 3b and 4b consist of checks for constructions 3 and 4,
combined with checks for condition 2 for strength t-1.
}

\examples{
checkcond1(3,2,11)  ## FALSE
checkcond2(3,2,11, progress=TRUE)  ## TRUE, example for progress bar
## if construction 4 with strength 4 works,
## then this setting will also fulfill construction 4b
checkcond4b(4,2,11, progress=TRUE)
N_cyc(11,2,"4b")
k_cyc(11,2,"4b")
## this array is the currently best-known
eCAN(4,12,2)

}
