% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CK_constrD.R
\name{CK_constrD}
\alias{CK_constrD}
\alias{CK_constrDCA}
\alias{N_CK_constrDCA}
\title{Function to augment a design with Chateauneuf-Kreher construction D}
\usage{
CK_constrDCA(k, v, ...)

CK_constrD(D3, D2=NULL, check=FALSE, start0=TRUE, ...)

N_CK_constrDCA(t=3, k, v, ...)
}
\arguments{
\item{D3}{a strength 3 CA with k columns in v-1 levels, coded as integers
0, ..., v-2 or 1,...,v-1}

\item{D2}{a strength 2 CA with k-1 columns in v-1 levels, coded as integers
0, ..., v-2 or 1,...,v-1, same coding as \code{D3};\cr
\code{NULL} implies automatic creation with function \code{\link{bestCA}},
which uses the known optimal function \code{\link{KSK}} for \code{v}=2,
and the best implemented construction for larger \code{v}.}

\item{check}{a logical: if TRUE, the strength requirements are verified
(may take very long for large arrays, therefore defaults to FALSE)}

\item{start0}{logical: Do integer values start with 0 ?  (\code{D3} and \code{D2}
must be compatible)}

\item{...}{further arguments to function \code{\link{coverage}} (\code{CK_constrD})
or \code{\link{bestN}} (function \code{N_CK_constrDCA})}

\item{k}{integer: the target number of columns}

\item{v}{integer: the number of levels}

\item{t}{integer: the target strength (always 3)}
}
\value{
Function \code{CK_constrDCA} returns a matrix of class \code{ca}
with attributes, which is a strength 3 covering array.\cr
Function \code{CK_constrD} returns a matrix without class and attributes,\cr
function \code{N_CK_constrDCA} returns the run size of the CA.

Note that the run size may depend on the availability of an internet connection
for cases for which the best strength 3 CA for \code{k} columns or the
best strength 2 CA for \code{k-1} columns is
in the Dwyer data base or the NIST library only
(these might exist - has not been checked).
}
\description{
A strength 3 CA in k (v-1)-level columns and N runs is extended
to k v-level columns in N + k\emph{M + k}(v-1) runs using a strength 2 CA in
M runs for k-1 (v-1)-level columns (construction D)
}
\details{
\code{CK_constrDCA} is the interface function that uses the workhorse
function \code{CK_constrD}.
}
\examples{
N_CK_constrDCA(3,6,6)
bestN(3,6,6) ## not competitive
dim(CK_constrD(SCA_Busht(5,3)))
dim(CK_constrDCA(6,6)) ## always strength 3

}
\references{
Chateauneuf and Kreher (2002)
}
