% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CK_NRB.R
\name{CK_NRB}
\alias{CK_NRB}
\alias{N_CK_NRB}
\alias{k_CK_NRB}
\title{Construction Chateauneuf-Kreher NRB of strength 3 for v=3,4,5 with k=2*v}
\usage{
CK_NRB(k, v)

N_CK_NRB(t=3, k, v)

k_CK_NRB(t=3, N, v)
}
\arguments{
\item{k}{integer, number of columns needed (<= 2*\code{v})}

\item{v}{integer, number of levels for each column}

\item{t}{integer, strength (must be 3)}

\item{N}{integer, maximal number of affordable runs}
}
\value{
\code{CK_NRB} returns a covering array of 33 (\code{v}=3),
88 (\code{v}=4) or 185 rows (\code{v}=5),\cr
\code{N_CK_NRB} returns a number for valid entries and \code{NA} otherwise.
}
\description{
Function for three very good strength 3 CAs with v=3,4,5 and k=2*v
from Chateauneuf, Colbourn and Kreher 1999 and Chateauneuf and Kreher 2002
}
\section{Details}{

CA(33,3,6,3) is from Figure 1 of Chateauneuf, Colbourn and Kreher (1999),\cr
CA(88,3,8,4) is from Example 2.3 of the paper with the ingredient M for
from Figure 2 (where typos in the last row of M had to be fixed).\cr
CA(185,3,10,5) is from Theorem 3.4 of Chateauneuf and Kreher (2002,
with AGL(2,q) as stated above Theorem 3.1 and the ingredient M in Figure 5 of the paper).
}

\examples{
dim(CK_NRB(6, 3))
attributes(CK_NRB(9,5))
Ns(3, 7, 4)  ## substantially better than other constructions
Ns(3, 9, 4) ## not possible
dim(bestCA(3,8,4))
k_CK_NRB(3, 100, 4) ## with up to 100 runs, 8 columns affordable

}
\references{
Chateauneuf, Colbourn and Kreher (1999), Chateauneuf and Kreher (2002)
}
