% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CKRS_augment.R
\name{CKRS_augment}
\alias{CKRS_augment}
\alias{CKRS_augmentCA}
\alias{N_CKRS_augmentCA}
\title{Function to augment a CA with Eqn (5) of CKRS}
\usage{
CKRS_augmentCA(t, k, v, ...)

CKRS_augment(D, start0=TRUE, ...)

N_CKRS_augmentCA(t, k, v, ...)
}
\arguments{
\item{D}{a strength t CA with k columns in v-1 levels, coded as integers
0, ..., v-2 (or 1, ..., v-1)}

\item{start0}{logical: Do integer values start with 0 ?}

\item{...}{further arguments to function \code{\link{bestN}}
(function \code{N_CKRS_augmentCA})}

\item{t}{integer: the target strength}

\item{k}{integer: the target number of columns}

\item{v}{integer: the number of levels}
}
\value{
Function \code{CKRS_augmentCA} returns a matrix of class \code{ca}
with attributes, which is a strength t covering array.\cr
Function \code{CKRS_augment} returns a matrix without class and attributes,\cr
function \code{N_CKRS_augmentCA} returns the run size of the CA.

Note that the run size may depend on the availability of an internet
connection for cases for which the best strength t CA for \code{k}
columns with v-1 levels is
in the Dwyer data base or the NIST library only
(these might exist - has not been checked).
}
\description{
increase the number of levels of a CA by 1 in a greedy fashion
}
\details{
\code{CKRS_augmentCA} is the interface function that uses the workhorse
function \code{CKRS_augment}.
}
\examples{
dim(CKRS_augment(SCA_Busht(5,3))) ## not competitive
bestN(3,6,6)

}
\references{
CKRS (2010)
}
