% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositCA.R
\name{compositCA}
\alias{compositCA}
\alias{N_compositCA}
\title{Obtain a CA for non-prime number of levels by composition}
\usage{
compositCA(t, k, v, ...)

N_compositCA(t, k, v, ...)
}
\arguments{
\item{t}{positive integer: requested strength}

\item{k}{positive integer: requested number of columns}

\item{v}{positive integer, non-prime: requested number of levels for each column}

\item{...}{currently not used}
}
\value{
Function \code{compositCA} yields a matrix of class \code{ca}), with levels from 0 to \code{v-1}.
}
\description{
Function to construct a strength t CA in k columns at v levels, v non-prime, from two strength t CAs in fewer levels
}
\section{Details}{

Function \code{compositCA} factorizes \code{v} into all variants of two factors,
picks the pair that produces the smallest run size, and obtains a CA in \code{v} levels
by composition from two CAs with those numbers of levels, using function \code{\link{crossCAs}}.

It uses function \code{\link{bestCA}} for obtaining the ingredient CAs. The result is rarely optimal.
}

\examples{
################################################
## compositCA
################################################
D <- compositCA(3, 12, 9)
dim(D)
Ns(3, 12, 9) ## not competitive

D <- compositCA(4, 12, 4)
dim(D)
Ns(4, 12, 4) ## best implemented
eCAN(4, 12, 4)  ## this has not been implemented

Ns(3, 4, 12) ## best (orthogonal array)

Ns(4, 24, 12) ## slightly worse than fuseBushtCA

Ns(4, 12, 8)  ## best implemented

Ns(4, 12, 24) ## much worse than fuseBushtCA,
              ## which almost yields eCAN

}
