% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cphfCA.R
\name{cphfCA}
\alias{cphfCA}
\alias{N_cphfCA}
\alias{k_cphfCA}
\alias{WCS}
\title{Making CA from CPHF}
\usage{
cphfCA(t, k, v, start0 = TRUE, ...)

N_cphfCA(t, k, v)

k_cphfCA(t, N, v)

WCS(cphf, q, t, type="2022", ...)
}
\arguments{
\item{t}{integer: target strength}

\item{k}{integer: requested number of columns}

\item{v}{integer: number of levels, must be a prime power}

\item{start0}{logical: should the levels of the output array start with 0 (otherwise with 1)?}

\item{...}{currently not used}

\item{N}{integer: affordable run size}

\item{cphf}{matrix with integer entries that can be expanded into permutation vectors of length \code{t}; must be a CPHF}

\item{q}{integer: number of levels, must be a prime power}

\item{type}{character string providing the type of construction; for the built-in CPHFs, \code{"2022"} is the correct entry.
It handles CPHFs for which the first entry of the permutation vector refers to the highest power,
and which were constructed with Galois fields from SageMath or the Python \code{numbthy} module
(relevant only for prime powers 8, 9, 16, 25; also implemented in this package via the internal function \code{mygf})\cr
Any type different from the default requests use of the built-in Galois fields of R package \pkg{\link[lhs]{lhs}}.}
}
\value{
Function \code{cphfCA} returns a matrix of class \code{ca} with attributes,
which is a covering array of strength 3, 4, 5 or 6.\cr
Function \code{WCS} returns a matrix with entries in 0,...,v-1, which is a covering
array of strength \code{t}, if a valid \code{cphf} for strength \code{t} is provided;
the function is not primarily meant for direct use,
but is the workhorse behind \code{cphfCA}.

Functions \code{N_cphfCA} and \code{k_cphfCA} return the required run size or the achievable number of columns, respectively.
}
\description{
according to the process outlined in Wagner, Colbourn and Simos (2022)
}
\section{Details}{

Function \code{cphfCA} implements CAs from the Wagner et al. (2022) construction
for length \code{t} h-tuples to create length \code{v^t} permutation
vectors for \code{v} levels and strength \code{t}.
The workhorse function \code{WCS} uses a stored covering perfect hash family (CPHF); at present,
only the CPHFs from Wagner, Colbourn and Simos (2022) are implemented via the function
\code{cphfCA} (see \code{\link{WCS_CPHFs}}
and \code{\link{CPHFcat}}).\cr
Users who have their own valid CPHF (not SCPHF)
that is not part of this package can use the function for creating a CA. Note that
the creation of a CA can take a long time for large cases; for example, creation of a
CA(501121, 4, 665, 17) from the relevant CPHF with 6 rows (\code{6*(17^4-1)+1=501121})
took approximately 140 minutes on the authors Windows computer with 32GB RAM
(which appeared to be the limiting factor). This package is not optimized to handle cases
as large as this.

The construction relies on a Galois field GF(\code{v}).
The CPHFs from Wagner et al. (2022) for \code{v=8}, \code{v=9}, \code{v=16} and \code{v=25}
require a different Galois field than otherwise used in this package; these deviating
Galois fields are created with the internal function \code{mygf}; function \code{WCS} takes
the argument \code{type}, with which these Galois fields are activated (per default), but can
also be deactivated in favor of the Galois fields of package \pkg{\link[lhs]{lhs}}
by specifying any type different from \code{"2022"}.\cr
The data frame \code{\link{CPHFcat}} holds the overview information on which constructions
are implemented for function \code{cphfCA}.
}

\examples{
# applying the construction for small CPHF from the Wagner et al. (2022) paper

# strength 3
head(CPHFcat[which(CPHFcat$v==4),])
## the CPHF for the first row
WCS_CPHFs[["3"]][["4"]][["71"]]
D <- cphfCA(3,71,4)
dim(D)
\donttest{
  coverage(D,3)
}
Ns(3,71,4)   ## size is current best but not eCAN

## the CPHF for the 91st row has flexible entries
CPHFcat[91,]
rowSums(is.na(WCS_CPHFs[["4"]][["3"]][["3504"]]))
## many flexible values in the last row,
## eCAN claims one row less for this construction
\donttest{
  ## creation of this CA is slow because of the many columns
  D <- cphfCA(4,3504,3)  ## about 3 minutes on author's machine
  dim(D)
}
Ns(4,3504,3)   ## by far the best implemented

# strength 6
Ns(6, 10, 3)  ## by far best implemented,
              ## slightly better array from group who does not share their arrays
dim(D <- bestCA(6,10,3))
coverage(D, 6)


}
\references{
Wagner, Colbourn and Simos (2022)
}
