% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossCAs.R
\name{crossCAs}
\alias{crossCAs}
\title{Cross two strength t CAs with k columns to yield strength t in v*w levels}
\usage{
crossCAs(A, B, t=NULL, check=FALSE, ...)
}
\arguments{
\item{A}{an N x k CA of strength \code{t} with v levels,}

\item{B}{an M x k CA of strength \code{t} with w levels,}

\item{t}{strength of \code{A} and \code{B}, needed for \code{check=TRUE}}

\item{check}{logical; if TRUE, checks required strength of ingoing CAs (may substantially increase run time for larger CAs, especially for \code{t}>2)}

\item{...}{currently not used}
}
\value{
Function \code{crossCAs} yields a CA(N\emph{M, t, k, v}w) (N*M x k matrix of class \code{ca}).\cr
The returned matrix has its levels coded like the ingoing matrices, i.e., integer starting with 0 or 1.
}
\description{
Function to construct a strength t CA in k columns at v*w levels from a strength t CA in k columns at v levels and a strength t CA in k columns at w levels
}
\section{Details}{

Function \code{crossCAs} yields a CA(N\emph{M, t, k, v}w) by combining the corresponding columns of the two CAs
into \code{v*w} levels, according to, e.g., Theorem 2.8 of Zhang et al. (2014).
}

\examples{
################################################
## crossCAs
################################################
## two CAs with four columns each and strength 3
A <- lhs::createBush(3, 4, bRandom=FALSE)  ## strength 3, 4 columns in 3 levels
dim(A)
B <- lhs::createBush(4, 4, bRandom=FALSE)   ## strength 3, 4 columns in 4 levels
dim(B)
## product array with 3*4=12 levels and 27*64=1728 runs
E <- crossCAs(A, B, 3)
dim(E)
tail(E)
coverage(E, 3)
eCAN(3, 4, 12)      ## the CA is optimal (as of April 2025)
eCAK(3, 1728, 12)   ## the optimal strength 3 12-level CA accommodates two more columns,
                    ## and even in an OA (12 is not a prime power, thus 13 columns not possible)

}
\references{
Chateauneuf and Kreher (2002), Zhang et al. (2014)
}
