% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CS_CMMSSY.R
\name{CS_CMMSSY}
\alias{CS_CMMSSY}
\alias{N_CS_CMMSSY}
\alias{k_CS_CMMSSY}
\title{Strength 2 CAs constructed by cover starters (CMMSSY)}
\usage{
CS_CMMSSY(k, v, start0=TRUE, starter=NULL, ...)

N_CS_CMMSSY(t=2, k, v, theoretical=FALSE)

k_CS_CMMSSY(N, v)
}
\arguments{
\item{k}{number of factors}

\item{v}{number of levels, \code{v} >= 3}

\item{start0}{should the output design's levels start with zero (default) or 1?}

\item{starter}{an optional cover starter vector that starts with \code{Inf}.\cr
This may be useful for cases with k larger than implemented in the stored cover starters.\cr
CAUTION: Responsibility for achieving the coverage rests with the user, please check!}

\item{...}{currently not used}

\item{t}{requested strength (has to be 2)}

\item{theoretical}{logical: default requests size available from using stored cover starters,\cr
change to \code{TRUE} for the size assuming existence of a cover starter for \code{k} and \code{v}}

\item{N}{affordable number of runs}
}
\value{
\code{CS_MMSSY} returns the smallest possible strength 2 CA from the Colbourn et al. (2006) construction for \code{k} experimental factors in \code{v} levels; it is an \code{N x k} matrix with entries from 0 to \code{v-1} or from 1 to \code{v}, depending on \code{start0}.\cr
\code{N_CS_MMSSY} returns the minimum run size needed for the CMMSSY construction with \code{k} factors at \code{v} levels, or \code{NA}, if no such construction exists.\cr
\code{k_CS_MMSSY} returns the maximum number of \code{v}-level factors that can be accommodated in at most \code{N} runs, or \code{NA}, if no such construction exists.
}
\description{
Function to construct a strength 2 CA based on cover starters according to Colbourn et al. 2006.
k and v must be specified, the number of runs is ((v-1)*k+1) (possibly for a larger value of k
if the construction does not exist for the requested k); recursive constructions based on such
arrays are also included
}
\section{Details}{

Colbourn et al. (2006) denote designs with the factors as the columns and the runs as the rows, like in this package, and contrary to Meagher and Stevens (2005).

The construction uses a distinct starter vectors of \code{k} elements with first element \code{Inf} (fixed position)
and subsequent elements from 1 to \code{v-1}.

The starter vector is cycled into a \code{k x k} matrix \code{M} (internal function \code{circular}),
which is one key ingredient of the construction. The other key ingredient is addition modulo \code{v-1}.
Finally, the infinite element is set to \code{v-1}.\cr

Functions \code{N_CS_CMMSSY} and \code{k_CS_MMSSY} return the minimum \code{N} or the maximum
\code{k} for the requested scenario. In case of \code{N_CS_MMSSY}, the default to use the
available starters can be overridden by the argument \code{theoretical}, which allows
to inspect how many runs can be obtained by a suitable user-specified starter vector.

Besides the starters from Colbourn et al. (2006), the function also returns arrays that
are obtained by combining pairs, triples or quadruples of such arrays,
using function \code{\link{productPCA}}, once (for pairs) or repeatedly.
It would be possible to drive this further, but the function stops at quadruples.
Such an approach was proposed for these distinct starters in Colbourn (2004), who also provided
a smaller selection of starter vectors. (The approach is analogous to \code{\link{recBoseCA}}.)
}

\examples{
## mini example of Colbourn et al. (2006)
k <- 7; v <- 5
starter <- c(Inf, 0, 0, 1, 3, 2)
CS_CMMSSY(7, 5, starter=starter) ## 29 runs, 7 columns

CS_CMMSSY(9, 6)  ## 46 rows

}
\references{
Colbourn (2004), Colbourn et al. (2006)
}
