% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclotomy.R
\name{cyc}
\alias{cyc}
\title{function for cyclotomy constructions}
\usage{
cyc(k, t, v, q, type = NULL, primitive = NULL)
}
\arguments{
\item{k}{number of columns}

\item{t}{requested strength (\code{t} < \code{k})}

\item{v}{number of levels for each column}

\item{q}{prime or prime power with q=1(mod v)}

\item{type}{the construction type ("1","2","3","3a","3b","4a", or "4b"),
default implies "1"}

\item{primitive}{the primitive of the finite field to be used for the construction
(specification should not be needed, except for cases that are not adequately covered
by the internal data.frame \code{primedat})}
}
\value{
a matrix with \code{k} columns and values 0 to \code{v-1} with coverage
strength \code{t} and the number of runs corresponding to the construction
(see Details section)
}
\description{
provides cyclotomy constructions according to Colbourn (2010), which yield
good high strength arrays for v=2,3,4,6 (see Details section)
}
\section{Details}{
 The function yields designs with columns in v levels based on
prime or prime power q as follows:

for N=k=q=1(mod v) prime power (Construction 4.1)\cr
or N=q+v-1, k=q, with q=1(mod v) prime power (Construction 4.2)\cr
(add constant rows with elements 1, 2, ..., v-1\cr
to construction 4.1, as these are missing)\cr
or N=vq, k=q, with q=1(mod v) prime power (Construction 4.3)\cr
(juxtapose vertically A, A+1, ..., A+v-1 (mod v);\cr
this is called developing modulo v)\cr
or 4.3a: k=q+1\cr
(extend A by column of 0s before developing)\cr
or 4.3b: N=vq+v^2-v, k=q+1\cr
(add non-zero constant rows (like in 4.2),\cr
then add column of 0s before developing)\cr
or N=vq+v, k=q, with q=1(mod v) prime power (Construction 4.4)\cr
add constant row of zeroes to A before developing\cr
or 4.4a: k=q+1\cr
also add constant column of zeroes\cr
or 4.4b: N=vq+v^2-v, k=q+1 (same as 3.b)\cr
add to A all nonzero constant rows\cr
and a column of ones, then develop
}

\examples{
## construction 1
## 2-level
CA19.3.19.2 <- cyc(19, 3, 2, 19)
coverage(CA19.3.19.2, 3)
coverage(CA19.3.19.2, 4)

# 3-level
CA62.3.32.3 <- cyc(32, 3, 3, 31, type="3a")
coverage(CA62.3.32.3, 3)
coverage(CA62.3.32.3, 4)

}
