% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse.R
\name{swapvals}
\alias{swapvals}
\alias{fuse}
\title{function to reduce the number of levels by 1}
\usage{
fuse(D, vprior, vpost, ...)
}
\arguments{
\item{D}{a CA with \code{vprior} levels in each column (integers, starting a 0 or 1)}

\item{vprior}{the number of levels prior to applying fuse}

\item{vpost}{the number of levels after applying fuse, must be smaller than vprior}

\item{...}{currently not used}
}
\value{
an array of the same coverage properties and the same number of columns
as \code{D} and at most \code{nrow(D) - 2 * (vprior - vpost)} rows
with \code{vpost} levels per column
}
\description{
reduces the number of levels by 1,
achieving a reduction of the number of rows by 2.
fuse can also reduce the number of levels by more than 1,
reducing the number of rows by 2 for each reduction of 1 level.
}
\section{Details}{

Note that there is a variant of fusion, implemented in function \code{\link{fuseBose}},
for which the number of runs can be reduced by three; this is applicable for strength 2
with \code{vprior} a prime or prime power and up to \code{vprior + 1} columns in \code{D}.
}

\examples{
set.seed(1212) ## for array randomization
D <- lhs::createBush(11, 12)
table(D)
dim(D)
D10 <- fuse(D, 11, 10) ## reduce by two rows and one level
table(D10)
dim(D10)
coverage(D, 3)
coverage(D10, 3)

}
\references{
Colbourn et al. (2010, CKRS; Lemma 3.1)
}
\seealso{
\code{\link[=fuseBose]{fuseBose()}} for the case mentioned in the Details section
}
