% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_needed.R
\name{identify_needed}
\alias{identify_needed}
\title{Identifying needed and redundant elements of SCAs}
\usage{
identify_needed(A, t, ...)
}
\arguments{
\item{A}{a sequence covering array (SCA) with k columns and elements 1,...,k in each row}

\item{t}{the strength for which the coverage of all permutations of length t must be guaranteed for all \code{combn(k,t)} t-element subsets of 1,...,k}

\item{...}{currently not used}
}
\value{
a logical matrix of the same dimensions as \code{A} with \code{TRUE} for needed values and \code{FALSE} for values that can be moved.
}
\description{
from a given SCA A and its strength t
}
\section{The meanings of TRUE and FALSE in the resulting matrix}{

In an SCA of strength t,
it is guaranteed that all orders of all t-tuples of elements of 1,...,k
are covered; the ordered tuples are also counted as covered if they are
interleaved by other values. For example, the sequence 1->4->3->2->5
covers the 3-sequence 1->3->5. The run 14325 covers \code{choose(5,3)}=10
3-sequences, and each of its elements is part of six of these.
If all six such sequences are already covered in earlier runs,
the element is marked as redundant (\code{FALSE}),
otherwise it is needed (\code{TRUE}); for example, if the sequences 1->4->3, 1->4->2, 1->4->5,
4->3->2, 4->3->5, 4->2->5 were already covered by earlier runs, the position of the "4" becomes \code{FALSE}.
Thus, if the position is \code{TRUE}, these sequences might still be covered
by later runs, but are not guaranteed to.
}

\section{Use of AI}{

Claude 4 was involved in the development of this function.
}

\examples{
### create a trivial SCA of strength 2
A <- rbind(1:5, 5:1)
identify_needed(A,2)

## a naive greedy SCA
A <- greedySCA_naive(5, 3, seed=2323, postop=FALSE)
dim(A)
## identify the necessary elements
(neededs <- identify_needed(A,3))
## there are various FALSE-only rows
## remove those
A <- A[!rowSums(neededs)==0,]
neededs <- neededs[!rowSums(neededs)==0,]
dim(A)
## all sequences are still covered
(cove <- coverageSCA(A, 3))
## the count_effective attribute of cove shows
## that the last three rows are only needed for 1 sequence each
attr(cove, "count_effective")
## correspondingly, two elements of each of those rows are redundant
neededs
A[11:13,]
## row 11 is needed for the sequence 3->1->4,
## row 12 for the sequence 5->4->3,
## and row 13 for the sequence 5->1->4

}
