% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSK.R
\name{N_KSK}
\alias{N_KSK}
\alias{k_KSK}
\title{function for optimal 2-level CAs of strength 2}
\usage{
KSK(k=NULL, N=NULL)

k_KSK(N)

N_KSK(k)
}
\arguments{
\item{k}{number of columns}

\item{N}{number of runs}
}
\value{
\code{KSK} returns a matrix with \code{k} columns and \code{N} runs and levels 0 and 1
in each column, with coverage strength 2.

\code{N_KSK} and \code{k_KSK} return the minimum \code{N} for a given \code{k}
or the maximum \code{k} for a given \code{N}, respectively.
}
\description{
implements the construction by Kleitman and Spencer (1973) and Katona (1973),
which have optimal size
}
\examples{
## smallest design for 10 columns
KSK(k=10)
## largest design for 6 rows
ca6.2.10.2 <- KSK(N=6)
coverage(ca6.2.10.2, 2)
coverage(ca6.2.10.2, 3)
## largest k for 10 rows
k_KSK(10)
## smallest N for 2800 columns
N_KSK(2800)
## design for 8 columns in 6 rows (not necessarily best subset of columns)
KSK(8,6)
## design for 2 columns (full factorial needed)
KSK(2)

}
