% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscCA.R
\name{miscCA}
\alias{miscCA}
\title{Incorporate miscellaneous arrays}
\usage{
miscCA(t, k, v, fixNA=TRUE, seed=NULL, maxconstant=FALSE, makePCA=FALSE, ...)
}
\arguments{
\item{t}{integer: the strength}

\item{k}{integer: the number of columns}

\item{v}{integer: the number of levels for each column}

\item{fixNA}{logical: should flexible values be fixed?\cr
If the array has flexible values, \code{fixNA=TRUE}
randomly assigns a fixed value to each flexible value.\cr
Thus, a seed is needed for reproducibility.}

\item{seed}{\code{NULL} or an integer seed for making the
fixing of NA values reproducible; a non-NULL seed
is stored in the attribute \code{fixNA_seed}
of the returned object.}

\item{maxconstant}{logical: should constant rows be maximized ?\cr
ignored for internal arrays, applied for arrays loaded
from package \pkg{\link[DoE.base]{DoE.base}} only;
per default, the array is not modified and may not even
have a single constant row}

\item{makePCA}{logical: should a PCA structure be enforced ?\cr
ignored for internal arrays, applied for arrays loaded
from package \pkg{\link[DoE.base]{DoE.base}} only\cr
If both \code{makePCA} and \code{maxconstant} are TRUE,
\code{maxconstant} takes precedence.}

\item{...}{currently not used}
}
\value{
an array of class \code{ca} with attributes
}
\description{
incorporates miscellaneous arrays that are created in the package
or obtained from package DoE.base
}
\section{Details}{

The function relies on the data frame \code{\link{miscCAcat}} for a
description of the available CAs.
}

\examples{
# a CA (without flexible values) from DoE.base
miscCA(3, 5, 4)

## without making rows constant and without enforcing PCA
attributes(miscCA(3, 5, 4, maxconstant=FALSE))

# a CA that is internally stored
# it has a single flexible value, which is per default fixed
# the seed makes the fixing reproducible
attributes(D <- miscCA(2, 14, 4, seed=374))
Ns(2,14,4)  ## best possible

}
\references{
Ji and Yin (2010), Colbourn and Torres-Jimenez (2013), Kuhfeld (w/o year),
\pkg{\link[DoE.base]{DoE.base}}
}
