% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hartman74.R
\name{N_Hartman74}
\alias{N_Hartman74}
\alias{Hartman74}
\alias{N_hartman74}
\alias{Turan}
\title{Construction in Hartman Theorem 7.4}
\usage{
Hartman74(A, v, t, B=NULL, ...)

N_Hartman74(N_A, k_A, v, t, ...)

Turan(t, v)
}
\arguments{
\item{N_A}{run size of the matrix \code{A} (called N above)}

\item{k_A}{number of columns of the matrix \code{A} (called k above)}

\item{v}{the number of levels (or the number of parts for \code{Turan})}

\item{t}{the strength (or the number of vertices for \code{Turan})}

\item{...}{currently not used}

\item{A}{an N x k CA of strength \code{t} columns in \code{v} levels each, denoted as 0 to v-1 or 1 to v}

\item{B}{NULL, or a strength 2 CA in at least \code{Turan(t,v)} columns with levels from 1 to k (levels from 0 to (k-1) are acceptable too). If not specified by the user, the Bose construction from \code{\link[lhs]{createBose}} is used, which yields at most k+1 columns, for a prime power k. If k is not a prime power or k+1 columns are not enough, the user must provide a suitable \code{B} or an \code{A} with more columns (may not be easy, at least not for achieving acceptable run sizes).}
}
\value{
function \code{Hartman74} returns a strength \code{t} CA with \code{k^2} columns,\cr
function \code{N_Hartman74} returns the achievable run size from using A, provided that a suitable B exists (warning, if this is not the case for the Bose construction for prime power k),\cr
and function \code{Turan} returns the maximum number of edges in a \code{v}-partite graph with \code{t} vertices.
}
\description{
The number of columns of an N x k design A can be squared; the design size is usually not optimal.
}
\section{Details}{

The construction vertically stacks \code{Turan(t,v)} copies of column selections from \code{A},\cr
where the ith copy is obtained by picking the \code{A} columns according to the elements of the ith column of \code{B}.\cr
Eventually, duplicate rows are removed.

The construction requires that the number of columns of \code{B} is at least \code{Turan(t,v)}, which may be impossible if \code{A} has too few columns. Sometimes, it can be possible to increase the number of columns for A (see examples).
}

\examples{
# Turan numbers (minimum number of columns needed for B)
Turanvec <- Vectorize(Turan)
TuranTable <- outer(1:13, 2:7, function(X, Y) Turanvec(X, Y))
dimnames(TuranTable) <- list(t=1:13, v=2:7)
TuranTable
## the first seven rows remain constant for larger values of v

A <- lhs::createBusht(4,5,3,bRandom=FALSE) ## 64 x 5
D <- Hartman74(A, v=4, t=3)
coverage(D, 3)
dim(D)
eCAN(3,25,4) ## D is far from optimal (roughly factor 1.6)

A <- lhs::createBusht(3,4,4,bRandom=FALSE) ## 81 x 4
A5 <- derive(DoE.base::L243.3.6-1)         ## 81 x 5
     ## optimal CA for strength 4 in 5 3-level columns according to Colbourn tables
# D <- Hartman74(A, v=3, t=4) ## would yield an error
D <- Hartman74(A5, v=3, t=4)
# coverage(D, 4)  ## commented out because of check time
dim(D)
eCAN(4,25,3) ## D is far from optimal (roughly factor 1.44)

}
\references{
Hartman (2005)
}
