% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_SCA.R
\name{simulated_annealing_sca}
\alias{simulated_annealing_sca}
\alias{optimize_SCA}
\title{Simulated Annealing for improving an SCA}
\usage{
optimize_SCA(A, t, sa_params = list(), target=factorial(t), sa=TRUE, verbose = FALSE, ...)

simulated_annealing_sca(A, t, maxAccepted = 100, maxAttempts = 1000, initialTemperature = 100, finalTemperature = 0.1, temperatureReduction = 0.95, verbose = FALSE, ...)
}
\arguments{
\item{A}{an SCA of strength \code{t}, or for \code{simulated_annealing_sca}, a partial SCA for which coverage is to be completed}

\item{t}{the strength for which the coverage of all permutations of length t must be guaranteed for all \code{combn(k,t)} t-element subsets of 1,...,k (k the number of columns of A)}

\item{maxAccepted}{positive integer}

\item{maxAttempts}{positive integer}

\item{initialTemperature}{positive number}

\item{finalTemperature}{positive number, less than initialTemperature}

\item{temperatureReduction}{positive number, less than 1; the closer to 1, the slower the temperature reduction, and the more iterations are run}

\item{verbose}{logical; if TRUE, printing progress information}

\item{...}{currently not used}

\item{sa}{logical, default TRUE; if FALSE, simulated annealing is suppressed}

\item{target}{number of rows at which no further improvement is attempted; defaults to the number of permutations needed to cover all orders for a single t-tuple, which is likely unachievable for most relevant problem sizes}

\item{sa_params}{list of parameters whose default values are to be changed; all parameters not in the list keep their default states}
}
\value{
in case of success, a strength t SCA with k columns is returned; all rows contain the integers 1 to k.
}
\description{
for k columns and strength t for the third stage of an algorithm by Torres-Jimenez et al. (2022)
}
\section{Details}{

The internal function \code{simulated_annealing_sca} implements the simulated annealing
approach of Torres-Jimenez et al. (2022) ??? check whether the perturbations are really the ones proposed there.
It is not meant to be applied directly but for being called by function
\code{optimize_SCA} for reducing the number of rows of an existing strength t SCA.
\code{optimize_SCA} implements a workflow for applying
\code{\link{reduce_rows_iterative_complete}} and subsequently applying simulated
annealing repeatedly.

Torres-Jimenez et al. (2022) describe a 3-stage approach, the third stage of which is
simulated annealing. Simulated annealing is invoked if postprocessing of an initial greedy SCA
is unable to find further rows that can be removed without deteriorating coverage.
The post-processing and simulated annealing are combined in function \code{optimze_SCA}.
While the simulated annealing is needed for
achieving the very good performance that the paper reports on,
it is often excessively slow at least in the implementation of this package.
Therefore, it can be suppressed by setting \code{sa=FALSE}.\cr
The paper claims that the method is most useful for larger problems,
but for these, the run time will likely be prohibitive.
}

\section{Use of AI}{

Claude 4 was heavily involved in the development of these functions,
by initially implementing the pseudo code from the paper.
}

\examples{
## a small SCA with 13 rows that is optimized quickly to 9
A_TJ <- greedySCA_TJ(6,3, postopt = TRUE, seed=2323)
dim(A_TJ)

A_TJ_reduced <- optimize_SCA(A_TJ, 3, target=9, verbose=TRUE)
A_TJ_redwoSimAnneal <- optimize_SCA(A_TJ, 3, sa=FALSE, verbose=TRUE)
coverageSCA(A_TJ_reduced, 3)
## one less row is possible, but requires a very slow simulated
## annealing optimization
\dontrun{
A_TJ_reduced8 <- optimize_SCA(A_TJ_reduced, 3, target=8, verbose=TRUE)
## without setting the target, one would have to wait until
## the simulated annealing for reducing to 7 runs failed
}

}
\references{
Torres-Jimenez et al. (2022)
}
\seealso{
\code{\link[=greedySCA_TJ]{greedySCA_TJ()}} for the greedy method by
Torres-Jimenez et al. (2022) and
\code{\link[=reduce_rows_iterative_complete]{reduce_rows_iterative_complete()}} for the
postoptimization stage method of that paper
}
\author{
Ulrike Groemping
}
