% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.coverage.R
\name{print.coverage}
\alias{print.coverage}
\title{Methods for class coverage}
\usage{
\method{print}{coverage}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{an object of class coverage, created by function coverage}

\item{...}{further arguments to print}

\item{verbose}{logical, if TRUE, prints the verbose content as well}
}
\description{
print method for class coverage
}
