% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_rows.R
\name{reduce_rows_complete}
\alias{reduce_rows_complete}
\alias{reduce_rows_iterative_complete}
\title{Reducing the number of rows for an SeqCA}
\usage{
reduce_rows_complete(A, t, verbose=FALSE, ...)

reduce_rows_iterative_complete(A, t, verbose=FALSE, ...)
}
\arguments{
\item{A}{a sequence covering array (SeqCA) with k columns and elements 1,...,k in each row}

\item{t}{the strength for which the coverage of all permutations of length t must be guaranteed for all \code{combn(k,t)} t-element subsets of 1,...,k}

\item{verbose}{logical; if \code{TRUE}, printed output is more detailed}

\item{...}{currently not used}
}
\value{
an SeqCA with attributes
}
\description{
by covering all t-orders that are missing after removing a certain row through
rearrangements of redundant elements in other rows
}
\section{Details}{

This can likely be made faster by first ordering the rows in decreasing order of the count_effective attribute of coverage.
}

\section{Use of AI}{

Claude 4 was heavily involved in the development of these functions.
}

\examples{
# a naive greedy SeqCA
A <- greedySeqCA_naive(6, 4, seed=17)
dim(A)  ## 67 rows
coverageSeqCA(A, 4) ## perfect coverage
\donttest{
  ## somewhat slow (about 60s on my machine)
  Areduced <- reduce_rows_iterative_complete(A, 4)
  dim(Areduced)  ## 42 runs, i.e. reduced by about 1/3
  coverageSeqCA(Areduced,4)   ## still perfect coverage
}

}
