% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqCA_Levenshtein.R
\name{SeqCA_Levenshtein}
\alias{SeqCA_Levenshtein}
\title{Levensthein creation of an optimal strength t SCA for k = t + 1 columns}
\usage{
SeqCA_Levenshtein(k, t = k - 1, a=0, ...)
}
\arguments{
\item{k}{positive integer, at least 3: the number of columns}

\item{t}{the strength for which the coverage of all permutations of length t must be guaranteed for all \code{combn(k,t)} t-element subsets of 1,...,k;
for this construction, \code{t = k - 1} is required.}

\item{a}{the partition class}

\item{...}{currently not used}
}
\value{
a strength t SeqCA with k columns; all rows contain the integers 1 to k. The SeqCA has t! runs,
i.e., optimal run size.
}
\description{
based on the very clear presentation in the master thesis by Na (2021)
}
\section{Details}{

The construction is based on partitioning the k! permutations of the values in \code{1:k}
according to the weights of their derivatives, where the derivative of a permutation p is obtainable
as \code{d <- (p[1:(k-1)] - p[2:k] > 0) + 0L}, and its weight is \code{w <- sum(d*1:t)}; partitioning
is w.r.t. to classes generated by the modulo k result of the weights (\code{w \%\% k}).
Each of the t + 1 parts is an SeqCA of strength t (Levenshtein 1991 as presented in Na 2021).

The function works fast for small k; as k! increases very fast (much faster than 2^k), k larger than 10 need
a lot of storage space and are also slow to process (k=10 on my machine takes about 26 seconds).
}

\examples{
# example code

## t=3, k=4
A0 <- SeqCA_Levenshtein(4, a=0)
A1 <- SeqCA_Levenshtein(4, a=1)
dim(A0); dim(A1)
coverageSeqCA(A0, 3)
coverageSeqCA(A1, 3)
## together, these cover half of the 4-sequences
coverageSeqCA(rbind(A0, A1), 4)

## t=6, k=7
dim(SeqCA_Levenshtein(7))

\donttest{
## runs about 26 seconds in R 4.3 on Windows 11
## with 32 GB RAM and Intel(R) Core(TM) i7-10610U CPU @ 1.80GHz (2.30 GHz)

## t=9, k=10
system.time(print(dim(SeqCA_Levenshtein(10))))

## t=10, k=11 will take more than 11 times as long
}

## t=15, k=16
try(SeqCA_Levenshtein(16))

## t=20, k=21
try(dim(SCA_Levenshtein(21)))

}
\references{
Levenstein (1991) and Na (2021)
}
\author{
Ulrike Groemping
}
