% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC.R
\name{smcCA}
\alias{smcCA}
\alias{SMC}
\alias{N_smcCA}
\alias{k_smcCA}
\title{Making CA from SCPHF}
\usage{
smcCA(t, k, v, start0 = TRUE, ...)

SMC(cphf, q, t, ...)

N_smcCA(t, k, v)

k_smcCA(t, N, v)
}
\arguments{
\item{t}{integer: target strength}

\item{k}{integer: requested number of columns}

\item{v}{integer: number of levels, must be a prime power}

\item{start0}{logical: should the levels of the output array start with 0 (otherwise with 1)?}

\item{...}{currently not used}

\item{cphf}{an SCPHF}

\item{q}{integer: prime power}

\item{N}{integer: affordable run size}
}
\value{
Function \code{smcCA} returns a matrix of class \code{ca} with attributes,
which is a covering array of strength 3 or 4.\cr
Function \code{SMC} returns a matrix with entries in 0,...,v-1, which is a covering
array of strength \code{t}, if a valid \code{cphf} for strength \code{t} is provided;
the function is not meant for direct use, but is the workhorse behind \code{smcCA}.

Functions \code{N_smcCA} and \code{k_smcCA} return the required run size or the achievable number of columns, respectively.
}
\description{
according to the process outlined in Sherwood, Martirosyan and Colbourn (2006).
It is assumed that the SCPHF refers to all non-reduced permutation vectors
or to all reduced permutation vectors, but not to a mix of both.
}
\section{Details}{

Function \code{SMC} uses a covering perfect hash family of a specific form (SCPHF)
that works with length \code{v^t} permutation vectors for \code{v} levels
and strength \code{t}. Function \code{smcCA} implements CAs from this
construction for the SCPHFs from Sherwood et al. (2006).
The construction relies on a Galois field GF(\code{v}).
The SCPHFs for \code{v=8} or \code{v=9} require a different Galois field
than otherwise used in this package (created with internal function \code{mygf}).

The data for the construction are in \code{\link{SMC_CPHFs}} and \code{\link{SMCcat}},
respectively.
}

\examples{
# applying the construction for small SCPHF from the Sherwood et al. (2006) paper

# strength 3
cphf <- rbind(0:15,
              c(0,1,4,5,2,3,6,7,9,8,13,12,11,10,15,14))
D <- SMC(cphf, q=4, t=3)
dim(D)
coverage(D,3)
eCAN(3,16,4)   ## size matches eCAN

# strength 4
cphf <- rbind(0:7,
              c(0, 1, 2, 4, 3, 6, 7, 5))
D <- SMC(cphf, q=2, t=4)
dim(D)
coverage(D,4)
eCAN(4,8,2)   ## size larger than eCAN

}
\references{
Sherwood, Martirosyan and Colbourn (2006)
}
