% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAEX.R
\name{CAEX}
\alias{CAEX}
\title{CAEX: current optimal CAs of strength 2 with v=3}
\usage{
CAEX(k=NULL, N=NULL, t=2, v=3, ...)
}
\arguments{
\item{k}{positive integer or \code{NULL}; the requested number of factors;\cr
if both \code{k} and \code{N} are specified, \code{k} takes precedence;\cr
if both \code{k} and \code{N} are \code{NULL}, function \code{CAEX} throws an error.}

\item{N}{positive integer or \code{NULL}; the requested number of runs; run sizes up to 50 are implemented;
these accommodate up to 112770(!) columns\cr}

\item{t}{integer, the requested strength; currently, only 2 is implemented}

\item{v}{integer, the requested number of levels; currently, only 3 is implemented}

\item{...}{currently not used}
}
\value{
Function \code{CAEX} returns a CA with levels coded from 0 to v-1, and flexible positions
denoted as \code{NA} and an attribute that shows the PCA status (with k1 and k2 arising from the product
without postprocessing in case of product constructions).\cr
}
\description{
Functions for CAEX CAs of strength 2 with v=3.
These yield current optimal CAs.
The functions use stored CAs from Torres-Jimenez, Acevedo-Juarez
and Avila-George (2021), where necessary,
and product constructions from that paper, where possible.
}
\section{Details}{

Function \code{CAEX} implements the arrays from Torres-Jimenez, Acevedo-Juarez and Avila-George (2021).
It uses internally-stored CAs (in object \code{CAEX_CAs}) where the last construction step is based on numerical column extension
(N=11 to 33 and 35), and product constructions as indicated in Table  of Torres-Jimenez,
Acevedo-Juarez and Avila-George (2021) for N = 34 and 36 to 50.
The information on how to proceed comes from the internal list \code{CAEX_lineages}.\cr
The arrays have values starting with 0; flexible values, if any, are denoted as \code{NA}.
Function \code{CAEX} provides only strength 2 CAS, and currently only for 3-level columns.
These are the current best ones, i.e., the nessecary run size for a desired number of columns \code{k} can be obtained from function \code{\link{eCAN}},
or from function \code{\link{N_TJcat}}.
}

\examples{
## a stored CA
CAEX(12) ## k=12 columns
eCAN(2,12,3)

## a CA obtained from other CAs without flexible values by function \code{\link{productCA1}}
D <- CAEX(N=34) ## should have N=34 runs but currently has 35
                ## --> needs improvement of function productCA1
attributes(D)
is.PCA(D)

## a CA obtained from other CAs with flexible values by function \code{\link{productCA1}}
## not run because of slightly longish run time
\dontrun{
D <- CAEX(N=43) ## should have N=43 runs, but has 44
                ## improvement to productCA1 needed
attributes(D)
}

}
\references{
Torres-Jimenez, Acevedo-Juarez and Avila-George (2021)
}
