% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAN.R
\name{CAN}
\alias{CAN}
\title{functions for N from Colbourn tables and NISTcat}
\usage{
CAN(t, k, v)

NISTcat(t, k, v)
}
\arguments{
\item{t}{coverage strength}

\item{k}{number of columns}

\item{v}{number of levels for each column}
}
\value{
\code{CAN} returns a data frame with the smallest known run size (CAN)
and the corresponding source entry (Source), \code{N_NISTcat} returns the
run size of the corresponding catalogued array of the NIST covering array tables.
In cases for which there is no entry in the Colbourn table or NIST catalogue,
the returned number is a missing value.
}
\description{
The functions extract the number of runs needed for a specified setting
according to the Colbourn tables and according to the NIST catalogues.
}
\examples{
CAN(3, 199, 2)
N_NISTcat(3, 199, 2)

}
