% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAN.R
\name{eCAN}
\alias{eCAN}
\alias{N_NISTcat}
\alias{N_TJcat}
\title{functions for N from Colbourn tables and other libraries}
\usage{
eCAN(t, k, v)

N_NISTcat(t, k, v)

N_TJcat(t, k, v)
}
\arguments{
\item{t}{coverage strength}

\item{k}{number of columns}

\item{v}{number of levels for each column}
}
\value{
\code{eCAN} returns a data frame with the smallest known run size
(empirical CAN, based on the Colbourn table)
and the corresponding source entry (Source),
\code{N_NISTcat} returns the run size of the
corresponding catalogued array of the NIST covering array library,
and \code{N_TJcat} the same thing for the small library by
Jorge Torres-Jimenez.
In cases for which there is no entry in the respective
table or library, the returned results are missing values.
}
\description{
The functions extract the number of runs needed for a specified setting
according to the Colbourn tables, according to the NIST library of CAs
and the Torres-Jimenez library of CAs.
}
\examples{
eCAN(3, 199, 2)
N_NISTcat(3, 199, 2)
N_TJcat(3, 199, 2) ## equals the best-known array

eCAN(4, 199, 2)
N_TJcat(4, 199, 2) ## Colbourn table outdated

}
