% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CK.R
\name{CK_constrD}
\alias{CK_constrD}
\title{Function to create a strength 3 CA with Chateauneuf-Kreher construction D}
\usage{
CK_constrD(D3, D2=NULL, verify=FALSE, start0=TRUE, ...)
}
\arguments{
\item{D3}{a strength 3 CA, integers 0, ..., v-1 or 1,...,v}

\item{D2}{a strength 2 CA; if NULL, creates it by function \code{\link{KSK}}
(only for 2-level CAs); must have the same coding as \code{D3};
for prime or prime power v with up to v+1 columns, the user might use a Bose array
that can be created with function \code{\link[lhs]{createBose}}.}

\item{verify}{a logical: if TRUE, the strength requirements are verified
(may take very long for large arrays, therefore defaults to FALSE)}

\item{start0}{logical: Do integer values start with 0 ?}

\item{...}{further arguments to function \code{\link{coverage}}}
}
\description{
A strength 3 CA in k v-level columns and N runs is extented
to 2k v-level columns in N + M*(v-1) runs
by adding cyclic permutations of a strength 2 CA in v levels
and M runs
}
\details{
Chateauneuf and Kreher's (2002) construction D does not necessarily yield smallest arrays.
}
\examples{
# create a CA(19,3,22,2) from a CA(12,3,11,2) and a CA(7,2,11,1)
A <- paleyHad(11)
B <- KSK(k=11)
CK_constrD(A, B)  ## explicitly specify B
CK_constrD(A)     ## this B is also used automatically (for v=2 only)
mistaken <- CK_constrD(A, B[-7,])  ## D2 does not have strength 2
coverage(mistaken, 3)      ## without verify, strength 3 is not guaranteed
# CK(A, B[-7,1], verify=TRUE)   ## would throw an informative error

A <- lhs::createBush(3, 4, bRandom=FALSE)  ## strength 3 OA
B <- lhs::createBose(3, 4, bRandom=FALSE)   ## strength 2 OA
CK_constrD(A, B)                ## 45 runs, 8 3-level columns
CK_constrD(A+1, B+1, start0=FALSE)      ## levels starting with 1
eCAN(3, 8, 3)           ## 42 runs are smallest known size

A <- lhs::createBush(5, 6, bRandom=FALSE)  ## strength 3 OA
B <- lhs::createBose(5, 6, bRandom=FALSE)   ## strength 2 OA
CK_constrD(A, B)                 ## 225 x 12
eCAN(3, 12, 5)           ## 225 runs are smallest known size
## the analogous strategy also yields the best-known strength 3 CAs
##    for v=7,8,9,11,13,16,17,19,23,25 with 2v+2 columns

}
