% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CK_doubling.R
\name{CK_doubling}
\alias{CK_doubling}
\alias{CK_doublingCA}
\alias{N_CK_doublingCA}
\title{Function to create a strength 3 CA with Chateauneuf-Kreher doubling}
\usage{
CK_doublingCA(k, v, ...)

CK_doubling(D3, D2=NULL, check=FALSE, start0=TRUE, ...)

N_CK_doublingCA(t=3, k, v, ...)
}
\arguments{
\item{D3}{a strength 3 CA with j columns in v levels, coded as integers
0, ..., v-1 or 1,...,v}

\item{D2}{a strength 2 CA with j columns in v levels, coded as integers
0, ..., v-1 or 1,...,v, same coding as \code{D3};\cr
\code{NULL} implies automatic creation with function \code{\link{bestCA}},
which uses the known optimal function \code{\link{KSK}} for \code{v}=2,
and the best implemented construction for larger \code{v}.}

\item{check}{a logical: if TRUE, the strength requirements are verified
(may take very long for large arrays, therefore defaults to FALSE)}

\item{start0}{logical: Do integer values start with 0 ?  (\code{D3} and \code{D2}
must be compatible)}

\item{...}{further arguments to function \code{\link{coverage}} (\code{CK_doubling})
or \code{\link{bestN}} (function \code{N_CK_doublingCA})}

\item{k}{integer: the target number of columns}

\item{v}{integer: the number of levels}

\item{t}{integer: the target strength (always 3)}
}
\value{
Function \code{CK_doublingCA} returns a matrix of class \code{ca}
with attributes, which is a strength 3 covering array.\cr
Function \code{CK_doubling} returns a matrix without class and attributes,\cr
function \code{N_CK_doublingCA} returns the run size of the CA.

Note that the run size may depend on the availability of an internet connection
for cases for which the best strength 3 CA for \code{ceiling(k/2)} columns is
in the Dwyer data base or the NIST library only
(these might exist - has not been checked).
}
\description{
A strength 3 CA in j v-level columns and N runs is extended
to k=2j v-level columns in N + M*(v-1) runs
by adding cyclic permutations of a strength 2 CA in v levels
and M runs (Thms 4.1 and 4.5)
}
\details{
\code{CK_doublingCA} is the interface function that uses the workhorse
function \code{CK_doubling}.

Chateauneuf and Kreher's (2002) doubling contributes a few best-known
strength 3 CAs to the Colbourn tables; in particular, where v is a prime
or prime power and thus fulfills the conditions for the Bose
construction of a strength 2 OA and the Bush
construction of a strength 3 OA (as implemented in \code{SCA_Busht}),
the designs are at present (13 Feb 2025) best-known for v >= 5.
}
\examples{
(E1 <- CK_doublingCA(8, 2))
coverage(E1, 3)  ## successful
eCAN(3,8,2)   ## not optimal, but close

## the above was achieved with these ingredients:
D3 <- bestCA(3, 4, 2)  ## 8x4, OA
dim(D3)
coverage(D3, 3)
D2 <- bestCA(2,4,2)   ## 5x4, KSK construction
dim(D2)
E2 <- CK_doubling(D3, D2)
dim(E2)
table(E1-E2) ## content is the same, attributes differ

D3 <- paleyCA(3, 11)
coverage(D3, 3)  # suitable
E <- CK_doubling(D3)
dim(E)
coverage(E, 3)  ## successful
eCAN(3,22,2)   ## optimal

E3 <- CK_doublingCA(12, 4)
dim(E3)
## Thm 4.5, 4 levels
## ingredients for E3:
D3 <- DoE.base::L64.4.6 - 1 ## 6 columns are not created from SCA_Busht
  ## bestCA(3, 6, 4) would yield a suitable CA from the DWYER database,
  ## provided that there is an internet connection
D2 <- bestCA(2, 6, 4)        ## 19 runs, optimal
eCAN(2, 6, 4)
## E3 has 64 + (4-1)*19 = 121 rows
eCAN(3, 12, 4)    ## size almost optimal

## not all outcomes are so close to optimal
dim(CK_doublingCA(16, 3))
## 42 + 2*13 = 68 runs
eCAN(3, 16, 3)  ## run size too large by factor 4/3 versus the current optimum 51

}
\references{
Chateauneuf and Kreher (2002)
}
