% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\alias{prepcoverplot}
\alias{coverplot}
\alias{coverplot_tuples}
\title{coverage and diagnostic plot functions}
\usage{
coverage(D, t, isInteger=TRUE, verbose=0, start0=TRUE, parallel=1)

coverplot(D, t, isInteger=TRUE, start0=TRUE,
    type="combinations",
    main=paste0("Coverage chart for t=",t),
    xlab="\% projections", ylab="\% covered",
    col=rgb(0,0,1,0.5), las=1, plot=TRUE, ...)
}
\arguments{
\item{D}{array}

\item{t}{small positive integer, interaction strength to be considered}

\item{isInteger}{logical (default \code{TRUE}): are the elements of D all
consecutive integers starting with 0 or 1?}

\item{verbose}{0 (default) returns only four percentages, 1 and 2 return more detail.
Anything apart from 1 and 2 behaves like 0.}

\item{start0}{logical (default \code{TRUE}): do the integers start with 0 ? (irrelevant for \code{!isInteger})}

\item{parallel}{number of threads to be used (default: 1)}

\item{type}{\code{"projections"} (default) or \code{"tuples"}}

\item{main}{title of plot}

\item{xlab, ylab}{axis titles}

\item{las}{orientation of tickmark labels}

\item{col}{color}

\item{plot}{logical (default \code{TRUE}): generate the plot?}

\item{...}{further arguments to polygon}
}
\value{
\code{coverage} returns a list with elements \code{total},
\code{ave}, \code{min} and \code{simple}, with increasing detail added for
verbose 1 or 2.

The \code{coverplot} functions return a list of \code{x}, \code{y}, \code{t}
and \code{proportions} (from call with \code{verbose=1}).
}
\description{
functions for calculating covering proportions for a given strength,
and for visualizing those
}
\section{Details}{
 The proportions calculated are the proportion of covered
\code{t}-tuples (\code{total}, taking the tuples as units), the average proportion of covered
\code{t}-tuples taken over the \code{t}-factor projections (\code{ave}),
the \code{min}imum proportion of \code{t}-tuples covered taken over the \code{t}-factor
projections, and the proportion of \code{t}-factor projections that are
completely covered (\code{simple}).

The plot either plots the percentage of projections on the horizontal axis and
the coverage on the vertical axis (visualizing the average taken over projections)
or the percentage of tuples (i.e., projections as weighted by tuples) on the
horizontal axis, visualizing the total percentage.
}

\examples{
## fixed levels
D <- KSK(k=10)
coverage(D, 2)
coverage(D, 3)
## mixed levels, total and ave are different
plan <- cbind(rep(0:2, each=3), c(rep(0:1, 4), 1), c(rep(0:1, each=4),0))
coverage(plan, 2)
coverage(plan, 2, verbose=1)
coverage(plan, 2, verbose=2)
## two of three projections fully covered,
## one covered 5/6 only
coverplot(plan, 2, type="projections")
## the not fully covered projection involves the 3-level
## column and thus has 6/(6+6+4)=37.5\\% of the tuples instead of 1/3
coverplot(plan, 2, type="tuples")

}
