% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataForCPHF.R
\docType{data}
\name{WCS_CPHFs}
\alias{WCS_CPHFs}
\alias{CPHFcat}
\title{Data for the CPHF-based (permutation vector) constructions}
\format{
\code{WCS_CPHFs} is a list of lists:

The first hierarchy level is the strength \code{t}:\cr
There are four lists for strengths 3 to 6,\cr
each of which holds lists for 13 implemented numbers of levels \code{v},
which in turn hold lists of the implemented numbers of columns \code{k}.

\code{CPHFcat} is a data frame with columns \code{t}, \code{v}, \code{k},
\code{nCPHF}, \code{hasNA} and \code{N}. The construction is from
Wagner, Colbourn and Simos (2022), with arrays provided by Michael Wagner.

An object of class \code{data.frame} with 292 rows and 6 columns.
}
\source{
Wagner, Colbourn and Simos (2022)
}
\usage{
WCS_CPHFs

CPHFcat
}
\description{
List WCS_CPHFs and data frame CPHFcat that support the CPHF-based constructions
}
\section{Details}{

For more details, see also the documentation of function
\code{\link{cphfCA}}.\cr
The creation of the objects is documented in the R file
\code{WCS_CPHFs_from_read.R} in the \code{extdata}
folder of this package, which can be located using
\code{system.file("extdata", package="CAs")};
the implemented raw CPHFs are not available in the package.
}

\examples{
## an impression of the content
head(CPHFcat)
tail(CPHFcat)
names(WCS_CPHFs[["6"]][["3"]])
WCS_CPHFs[["6"]][["3"]][["7"]]  ## only one row --> CA with 3^6 runs
dim(WCS_CPHFs[["6"]][["3"]][["36"]])  ## 10 rows --> CA with 10*(3^6-1) + 1 runs

}
\references{
Wagner, Colbourn and Simos (2022)
}
\keyword{datasets}
