% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataForpowerCA.R
\docType{data}
\name{powerCTcat}
\alias{powerCTcat}
\title{Data for function powerCA}
\format{
\code{powerCTcat} is a data.frame with columns \code{t} (strength),
\code{v} (number of levels), \code{k} (number of CA columns),
\code{N} (number of CA rows), \code{Source} (the Source entry in
\code{\link{colbournBigFrame}}),
\code{Pbase}, \code{expon} (the base with its exponent and the resulting
OA for the construction of the DH(H)F), \code{M} (the number of rows of the DHHF),
\code{w1} (the number of levels of the first group of \code{u1} rows of the DHHF,
corresponding to the number of columns of the first CA to be constructed
with construction \code{constr1}, whose number of rows is \code{N1} and whose
number of constant rows is (at least) \code{c1}, ... (same for 2, 3, 4, where applicable),
\code{chi} the number of rows to re-add into the final CA for covering constant t-tuples),
\code{claimedN} and \code{claimedk} (the corresponding \code{N} and
\code{k} entries in \code{\link{colbournBigFrame}}).
}
\source{
The source are the Colbourn tables, combined with
Colbourn and Torres-Jimenez (2010).\cr
The data have been compiled based on the data.frame \code{colbournBigFrame}
that holds the information from the Colbourn tables, and based on the paper by
Colbourn and Torres-Jimenez (2010) for the construction. The code can be found
the \code{extdata} folder of this package, which can be located using
\code{system.file("extdata", package="CAs")}.
}
\usage{
powerCTcat
}
\description{
The data.frame powerCTcat holds the details for power constructions
according to Colbourn and Torres-Jimenez 2010
}
\section{Details}{

The data.frame \code{powerCTcat}
holds information on ingredients of the power construction according to
Colbourn and Torres-Jimenez (2010). It is used by the functions \code{\link{powerCA}},
\code{\link{N_powerCT}} and \code{\link{k_powerCT}}.

The columns \code{constr1} to \code{constr4} hold labels of the constructions
used for creating the ingredients; these were determined using function \code{\link{bestN}}.
In some cases, they may change in the future due to availability of improved constructions,
which will then also impact the achievable overall run size.
Users who spot such opportunities are invited to give a shout.

The columns \code{claimedN} and \code{claimedk} contain the \code{\link{eCAN}}
and \code{\link{eCAK}} values from the Colbourn tables in \code{\link{colbournBigFrame}};\cr
the columns \code{N} and \code{k} contain the run size and number of columns of actual
constructions based on the CAs that are implemented in this package.\cr
Implemented \code{N} are equal to \code{claimedN} for strengths 3 and 4,
within about 10\% of the optimum for strength 5 and a bit more spread-out for strength 6
(note that \code{claimedN} can only be reasonably compared to \code{N}, where \code{claimedk}
equals \code{k}).

Constructions with \code{claimedN>1000000} have been omitted with version 0.19;
this greatly improves maintenance and is not a practical limitation, as the large CAs
are likely not of value and cannot be created with this package by a typical computer.\cr
}

\examples{
## an impression of the content
xtabs( ~ t + v, data=powerCTcat)
xtabs( ~ t, data=powerCTcat)
boxplot(N / claimedN ~ t, data=powerCTcat,
         horizontal = TRUE, las=1)
xtabs( ~ t, data=powerCTcat, subset=claimedN<69000)
boxplot(N / claimedN ~ t, data=powerCTcat, subset=claimedN<69000,
         horizontal = TRUE, las=1)
xtabs( ~ t, data=powerCTcat, subset=claimedN<73000)
boxplot(N / claimedN ~ t, data=powerCTcat, subset=claimedN<73000,
         horizontal = TRUE, las=1)

with(powerCTcat, data.frame(t, v, k, N, claimedN, diff=(N-claimedN), Source)[claimedN>N,])

}
\references{
Colbourn and Torres-Jimenez (2010), Colbourn (wo year)
}
\keyword{datasets}
