% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataForSCPHF.R
\docType{data}
\name{SMC_CPHFs}
\alias{SMC_CPHFs}
\alias{SCPHFcat}
\alias{CL_SCPHFs}
\title{Data for the SCPHF-based (permutation vector) constructions}
\format{
\code{SMC_CPHFs} and \code{CL_SCPHFs} are lists of lists:

The first hierarchy level is the strength \code{t}:\cr
Both lists hold two lists for strengths 3 and 4,\cr
each of which holds lists for the implemented numbers of levels \code{v},
which in turn hold lists of the implemented numbers of columns \code{k}.

\code{SCPHFcat} is a data frame with columns \code{t}, \code{v}, \code{k}, \code{N}
and \code{type}. The latter indicates the source: for "2006", the source is
the Sherwood et al. (2006) paper and SCPHFs are in \code{SMC_CPHFs},\cr
for "2018" the source is work by Erin Lanus and co-authors,
and SCPHFs are in \code{CL_CPHFs} (arrays found in the Dwyer (2024) GitHub repository).

An object of class \code{data.frame} with 43 rows and 5 columns.

An object of class \code{list} of length 2.
}
\source{
Reference Sherwood, Martirosyan, Colbourn (2006), CPHFs and constructions
in the paper; Colbourn and Lanus (2018) and/or Colbourn, Lanus and Sarkar (2018),
SCPHFs from the repository of Dwyer (2024)
}
\usage{
SMC_CPHFs

SCPHFcat

CL_SCPHFs
}
\description{
Lists SMC_CPHFs and CL_SCPHFs and data frame SCPHFcat that support the SCPHF-based constructions
}
\section{Details}{

For more details, see also the documentation of function
\code{\link{scphfCA}}.\cr
The creation of the objects is documented in the R files \code{SherwoodMartirosyanColbournCPHF.R},
\code{Lanus_SCPHFs_from_DwyerRepo.R} and \code{SCPHF_dimchecks.R} in the \code{extdata}
folder of this package, which can be located using
\code{system.file("extdata", package="CAs")}; the implemented raw SCPHFs from the DWYER
repo are in the sub folder \code{SCPHFs} of that folder.
}

\examples{
## an impression of the content
SCPHFcat
SMC_CPHFs[["4"]][["3"]]

}
\references{
Colbourn and Lanus (2018), Colbourn, Lanus and Sarkar (2018), Dwyer (2024), Sherwood, Martirosyan, Colbourn (2006)
}
\keyword{datasets}
