% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUsefulArrays.R
\docType{data}
\name{oa1728.12.6}
\alias{oa1728.12.6}
\title{Useful ingredient arrays for covering array constructions}
\format{
\code{oa1728.12.6} is an orthogonal array (class \code{oa})
and a 1728 x 6 matrix with elements 0, 1, ..., 12. The array has strength 3.
it will eventually reside in the R package \pkg{\link{DoE.base}}.
}
\source{
The array was constructed by the package author using construction 2
of Ji and Yin (2010).
}
\usage{
oa1728.12.6
}
\description{
lists useful single arrays that are used as ingredients
for covering array constructions
}
\examples{
attributes(oa1728.12.6)

}
\references{
Ji and Yin (2010)
}
\keyword{datasets}
