% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHHF2CA.R
\name{DHHF2CA}
\alias{DHHF2CA}
\title{Make a larger uniform CA from a DHHF and smaller CAs}
\usage{
DHHF2CA(P, Dlist, v = max(Dlist[[1]]) + 1, ...)
}
\arguments{
\item{P}{an M x k DHHF of strength t for partition size p=min(t,v) with several blocks of rows,
where the ith block has w_i levels
(a PHHF is the case with partition size t)}

\item{Dlist}{a list of N_i x w_i CAs of strength t with \code{v} levels each}

\item{v}{number of levels of all elements of \code{Dlist},
automatically detected in case of coding as 0,...,\code{v}-1}

\item{...}{further arguments to function \code{\link{maxconstant}},
e.g., \code{one_is_enough=TRUE} for suppressing the search for more constant rows}
}
\value{
a CA(chi + sum_i=1^c (N_i-rho_i)*u_i, t, k, v)
(matrix, not of class \code{ca}, as there are no
guarantees regarding the strength, because guarding this is left to the user or to calling functions).
}
\description{
Function to construct a larger CA by replacing levels of a DHHF with columns of smaller CAs
}
\section{Details}{

A PHHF is a perfect heterogeneous hash family, a DHHF a distributed heterogeneous hash family
(weaker condition); heterogeneity means that there are c blocks with different numbers of levels,
w_1,...,w_c, and there are u_i rows for the ith block, with u_1 +...+ u_c = M.\cr
Function \code{DHHF2CA} yields a CA(chi + sum_i=1^c (N_i-rho_i)\emph{u_i, t, k, v) from
replacing the levels of the DHHF with the columns of the corresponding CA from
\code{Dlist},\cr
leaving out the rho_i constant rows of the ith CA and adding a few (chi) constant rows back
in, if needed, where chi = max(0, v − sum_i=1^c u_i}(v - rho_i)).
The construction is from Theorem 2.3 of Colbourn and Torres-Jimenez (2010).\cr
The function uses function \code{maxconstant} to maximize the constant rows of the CAs.

There may be (likely rare) situations with many different numbers of levels in the DHHF
for which the function's handling of constant rows could be slightly improved;
for these cases, the result will not have the minimum number of rows possible from
the construction, but one or very few more rows.

The function does not check for the properties of the ingoing arrays. Hence, it is the users
responsibility to ensure that these are as needed.
}

\examples{
#########################################
## DHHF2CA
#########################################

## the Ji and Yin (2010) strength 3 OA in six 12-level columns (Lemma 3.5 with Corollary 3.3)
## available in this package as oa1728.12.6
P <- t(oa1728.12.6[,1:5])
P <- Tred(P, 4:5, 1)
## two columns reduced to 11 levels each
dim(P)
eCAN(3, 12, 2) ## 15 runs, in TJ2level_CAs
eCAN(3, 11, 2) ## 12 runs, paleyCA also yields this
D <- DHHF2CA(P,
    list(bestCA(3,12,2), bestCA(3,11,2)))
dim(D)
eCAN(3, 1452, 2)  ## optimal, this construction


}
\references{
Colbourn and Torres-Jimenez (2010), Ji and Yin (2010) for the example
}
