% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedySeqCA_Kuhn.R
\name{greedySeqCA_Kuhn}
\alias{greedySeqCA_Kuhn}
\alias{iter_greedySeqCA_Kuhn}
\title{Greedy random creation of an SeqCA}
\usage{
greedySeqCA_Kuhn(k, t, nsamp=10, postopt=TRUE, seed=NULL, ...)

iter_greedySeqCA_Kuhn(niter, k, t, nsamp=10, postopt=FALSE,
    seed_init=NULL, verbose=TRUE, ...)
}
\arguments{
\item{k}{positive integer (at least t); the number of columns}

\item{t}{the strength for which the coverage of all permutations of length t must be guaranteed for all \code{combn(k,t)} t-element subsets of 1,...,k}

\item{nsamp}{the number of runs sampled in each step of the algorithm (the N of the paper); the default 10 will likely be too small for realistic applications}

\item{postopt}{logical; default depends on function; if TRUE, the function \code{\link{reduce_rows_iterative_complete}} tries to make rows redundant and remove them. If this takes too long, \code{postopt} should be FALSE.}

\item{seed}{an optional positive integer seed for random sampling; specifying a seed makes the array generation reproducible}

\item{...}{currently not used}

\item{niter}{the number of independent attempts to make}

\item{seed_init}{an optional positive integer seed for the first iteration; specifying a seed makes the array generation reproducible}

\item{verbose}{logical (default TRUE), controlling how much printed output is created}
}
\value{
a strength t SeqCA with k columns; all rows contain the integers 1 to k.
}
\description{
for k columns and strength t according to an algorithm by Kuhn et al. (2012)
}
\section{Details}{

The function \code{greedySeqCA_Kuhn} implements the construction of Kuhn et al. (2012). It starts
from t! random permutations of the the integers 1,...,k and adds a further random permutation,
until all orders of all t-tuples are covered.\cr
For \code{postopt=TRUE}, it then \emph{identifies the redundant elements and removes all rows that
contain redundant elements only (would be easiest, but at present not done)} applies iterative run size reduction
using function \code{reduce_rows_iterative_complete}.

The function \code{iter_greedySeqCA_Kuhn} iteratively repeats the process and keeps the best
outcome.
}

\section{Use of AI}{

Claude 4 was involved in the development of these functions (mainly for debugging).
}

\examples{
## without post-optimization
nrow(greedySeqCA_Kuhn(5, 3, seed=2323, postopt=FALSE))
## relatively unlucky
nrow(greedySeqCA_Kuhn(5, 3, seed=2415, postopt=FALSE))
## relatively lucky
nrow(A <- greedySeqCA_Kuhn(5, 3, seed=16, postopt=FALSE))
coverageSeqCA(A, 3)
coverageSeqCA(A, 4)

## with post-optimization
## differences become smaller (still relevant esp. for larger cases)
nrow(greedySeqCA_Kuhn(5, 3, seed=2323))
## relatively unlucky
nrow(greedySeqCA_Kuhn(5, 3, seed=2415))
## relatively lucky
nrow(A <- greedySeqCA_Kuhn(5, 3, seed=16))
coverageSeqCA(A, 3)
coverageSeqCA(A, 4) ## worse coverage of 4-sequences than for larger SeqCA

## iteration
nrow(iter_greedySeqCA_Kuhn(10, 5, 3, seed_init=2415)) ## start unlucky
  ## larger niter brings the run size down to 8
}
\references{
Kuhn et al. (2012)
}
\seealso{
\code{\link[=greedySeqCA_naive]{greedySeqCA_naive()}} for a naive random assembly of an SeqCA and \code{\link[=greedySeqCA_TJ]{greedySeqCA_TJ()}} for the graph-based greedy stage method of Torres-Jimenez et al. (2022)
}
\author{
Ulrike Groemping
}
