% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juxtapose.R
\name{juxtapose}
\alias{juxtapose}
\title{Function to juxtapose two strength t CAs}
\usage{
juxtapose(D1, D2, jcol=NULL, start0=TRUE, ...)
}
\arguments{
\item{D1}{a CA with levels as integers 0, ..., v-1 or 1,...,v}

\item{D2}{a CA like \code{D1}, same number of columns, same numbers of levels,
except perhaps for the \code{jcol}th column; must have the same coding as \code{D3}}

\item{jcol}{the id of the column to be concatenated; if \code{NULL},
it is either the only column id for which the numbers of levels differ, or 1.}

\item{start0}{logical: Do integer valued levels start with 0 ? (Otherwise, they must start with 1.)}

\item{...}{further arguments to function \code{\link{coverage}}}
}
\description{
Two CAs D1 and D2 with the same strength t, the same number of columns k,
and the same numbers of levels in all but at most one column
are juxtaposed, such that the levels of one corresponding column
in each array are concatenated, resulting in a new strength t CA
with the sum of the numbers of levels for that column.
}
\details{
Juxtaposing two CAs of a particular strength yields a new CA of the same strength.
If \code{D1} and \code{D2} do not have the same strength, the weaker strength prevails.
}
\examples{
# create a CA(24,3,2^10 4^1) from two CA(12,3,11,2)

## do it by hand
A <- paley(11)
A2 <- A
A2[,11] <- A[,11] + 2
coverage(rbind(A,A2),3)

## use the function
juxtapose(A, A, jcol=11)

## create a CA(13, 2^10 3^1, 2) by juxtaposing two
##    strength 2 CAs (one with a column that has only 1 level):
juxtapose(KSK(11), cbind(KSK(10), 0))
   ## note: it is not necessary to specify jcol
   ##       because there is only one column with
   ##       different number of levels

}
