% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Martirosyan.R
\name{Martirosyan4}
\alias{Martirosyan4}
\alias{Martirosyan5}
\title{Functions to construct a strength 4 or strength 5 CA from several input CAs}
\usage{
Martirosyan4(A, B, C, D, start0=TRUE, verify=FALSE, ...)

Martirosyan5(A, B, C, D, start0=TRUE, verify=FALSE, ...)
}
\arguments{
\item{A}{a strength 4 or 5 CA in k columns at v levels (0, ..., v-1 or 1, ..., v);
for k<=3, A can be smaller (see Details section).}

\item{B}{a strength 3 or 4 CA in k columns at v levels (0, ..., v-1 or 1, ..., v)}

\item{C}{a strength 2 or 3 CA in k columns at v levels (0, ..., v-1 or 1, ..., v);
constant rows can be omitted from C, and at least one rows can always be made
constant by permuting levels of some columns.}

\item{D}{a strength 2 CA in v columns at v levels (0, ..., v-1 or 1, ..., v).
For v=2, D can be chosen smaller (see Details section).}

\item{start0}{logical: do the values start at 0? This must hold for all input arrays.}

\item{verify}{logical: should the properties of input matrices be verified?
(this may take a long time for larger arrays)}

\item{...}{further arguments to function \code{\link{coverage}}, in case
\code{verify} is \code{TRUE} (esp., \code{parallel argument})}
}
\value{
a strength 4 CA with 2k columns at v levels each (same coding as ingoing CAs)
in \code{N_A + (v-1)*N_B + 2*N_D*N_C} rows, where N_array denotes the run
sizes of the arrays
}
\description{
based on Martirosyan and Trung (2004). The functions obtain a strength 4
or strength 5 CA (strength 5 not yet working)
in 2k columns at v levels each from several smaller arrays.
}
\details{
For \code{Martirosyan4}: The matrices need the lower of the two strengths, i.e., 4,3,2,2 for A,B,C,D.
C and D can in some cases be chosen smaller than CAs:
One can omit constant rows from \code{C} (and one can recode it to make at least one row constant),
and for v=2, one can choose matrix
\code{D} as 2x2 with constant rows (one row for each level). For k<=3, \code{A} can be chosen identical to \code{B}.

For \code{Martirosyan5}: The matrices need the higher of the two strengths, i.e., 5,4,3,2 for A,B,C,D.

Unless \code{verify} is TRUE, the user is responsible for providing suitable matrices. If the matrices are not adequate, the resulting array may be worse than expected.
}
\section{Warning}{
 At present, \code{Martirosyan5} does not yield strength 5, but only almost strength 5,
the root cause has not yet been found. Martirosyan and Trung (2004) have a typo in function h, which should
have an obvious correction (replace first or with and). Further mistakes in the paper or the code
have so far not been found.
}

\examples{
  A <- lhs::createBusht(5,6,4, bRandom=FALSE)
  B <- lhs::createBusht(5,6,3, bRandom=FALSE)
  C <- lhs::createBusht(5,6,2, bRandom=FALSE)
  D <- C[,-6]  ## need only v=5 columns
  C <- C[-1,]  ## constant first row of C omitted
  E <- Martirosyan4(A, B, C, D)
  coverage(E, 4)
  dim(E)
  eCAN(4, 12, 5)  ## E is far from optimal

  A <- lhs::createBusht(4,5,5, bRandom=FALSE)
  B <- lhs::createBusht(4,5,4, bRandom=FALSE)
  C <- lhs::createBusht(4,5,3, bRandom=FALSE)
  D <- lhs::createBusht(4,4,2, bRandom=FALSE)
  E <- Martirosyan5(A, B, C, D)
  coverage(E, 5)  ## something wrong
                  ## mistake in paper?
                  ## bug in programming?
  dim(E)
  eCAN(4, 12, 5)  ## run size of E is far from optimal,
                  ## even though not strength 5

}
