% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.coverage.R
\name{print.CAcoverage}
\alias{print.CAcoverage}
\title{Methods for class CAcoverage}
\usage{
\method{print}{CAcoverage}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{an object of class \code{CAcoverage}, created by function \code{\link{coverage}} or \code{coverage_iter}}

\item{...}{further arguments to print}

\item{verbose}{logical, if TRUE, prints the verbose content as well}
}
\description{
print method for class CAcoverage
}
