% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productCA.R
\name{productCA1}
\alias{productCA1}
\alias{productCA2}
\title{Product constructions}
\usage{
productCA1(D1, D2, check=TRUE, ...)

productCA2(A, B, t, check=FALSE, ...)
}
\arguments{
\item{D1}{an N x k CA of strength 2 with v levels}

\item{D2}{an N x l CA of strength 2 with v levels}

\item{check}{logical, if TRUE, checks required strength of ingoing CAs (may substantially increase run time for larger CAs, especially for \code{t}>2)}

\item{...}{currently not used}

\item{A}{an N x k CA of strength \code{t} with v levels,}

\item{B}{an M x k CA of strength \code{t} with w levels,}

\item{t}{strength of \code{A} and \code{B}}
}
\value{
Function \code{productCA1} returns a CA(N+M, 2, k\emph{l, v) (N+M x k}l matrix).\cr
Function \code{productCA2} yields a CA(N\emph{M, t, k, v}w) (N*M x k matrix).\cr
The returned matrices have their levels coded as the ingoing matrices, i.e., integers starting with 0 or 1.
}
\description{
Functions to construct a CA from a product construction of two ingoing CAs
}
\section{Details}{

\bold{I don't like the function names, they will likely change.}

Function \code{productCA1} yields a CA(N+M, 2, kl, v) from a product construction, as, e.g., described in Torres-Jimènez et al. (2019).\cr
Function \code{productCA2} yields a CA(N\emph{M, t, k, v}w) by combining the corresponding columns of the two CAs into \code{v*w} levels, according to Theorem 3.2 of Chateauneuf, Colbourn and Kreher (.\cr
}

\examples{
# product CA (N+M x k*l matrix)
A <- cyc(19,2)
B <- KSK(k=12)
dim(A); dim(B)
E <- productCA1(A, B)
dim(E)
coverage(E, 2)
## more than twice as large as optimal
eCAN(2,228,2)

## two CAs with four columns each and strength 3
A <- lhs::createBush(3, 4, bRandom=FALSE)  ## strength 3, 4 columns in 3 levels
dim(A)
B <- lhs::createBush(4, 4, bRandom=FALSE)   ## strength 3, 4 columns in 4 levels
dim(B)
## product array with 3*4=12 levels and 27*64=1728 runs
E <- productCA2(A, B, 3)
dim(E)
tail(E)
coverage(E,3)
eCAN(3,4,12)   ## the CA is optimal (as of April 2025)


}
\references{
Torres-Jimenez et al. (2019), Chateauneuf and Kreher (2002)
}
