% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCA.R
\name{readCA}
\alias{readCA}
\title{Function for reading a covering array from a text file}
\usage{
readCA(path, fn, flexible.symbols=c("*","-","."), comment.symbol="C",
origin=NULL, ...)
}
\arguments{
\item{path}{a character string that specifies the path}

\item{fn}{a character string that specifies the file name; see Section Details for requirements the file must satisfy.}

\item{flexible.symbols}{characters to be treated as flexible values (will be set to NA)}

\item{comment.symbol}{character that starts a comment line}

\item{origin}{character string to be attached to the output array as origin-attribute;
if \code{NULL}, the path and filename information is used}

\item{...}{further arguments for function \code{\link{read.table}},
which is used separately for each line (after reading them with \code{\link{readLines}})}
}
\value{
The output object is a matrix of class \code{ca}.
}
\description{
reads a text file into a matrix of class ca (for covering array)
}
\section{Details}{

The file \code{fn} must start with comment lines that begin with a comment symbol.\cr
The first line after the comments lines contains instructions about the file content,
in the form N k v1^k1 v2^k2 ... It is permissible for the instructions line
to have a single string without blanks or \code{^} (e.g., for the strength)
after the exponential notation for the columns; this will be ignored in the
reading process.\cr
All columns of the file must have the same starting value, i.e., start all with 0 or all with 1.
}

